+++ apps/server/src/database/prisma.ts
@@ -1,0 +1,26 @@
+import "dotenv/config";
+import { env } from "node:process";
+import { PrismaPg } from "@prisma/adapter-pg";
+import { PrismaClient } from "@prisma/client/extension";
+
+const databaseUrl = env.DATABASE_URL;
+
+if (!databaseUrl) {
+  throw new Error("DATABASE_URL is not defined.");
+}
+
+const adapter = new PrismaPg({ connectionString: databaseUrl });
+
+const globalForPrisma = globalThis as unknown as {
+  prisma?: PrismaClient;
+}; 
+
+export const prisma =
+  globalForPrisma.prisma ??
+  new PrismaClient({
+    adapter,
+  });
+
+if (env.NODE_ENV !== "production") {
+  globalForPrisma.prisma = prisma;
+}