+++ CHANGELOG.md
@@ -8,3 +8,5 @@ The Job Portal App Changes
   [FEATURE] Project: Configuration and Initial Setup
 - [JBM-6](https://shahidpj06.atlassian.net/browse/JBM-6)
   [FEATURE] App: Build the UI with Mock Data
+- [JBM-7](https://shahidpj06.atlassian.net/browse/JBM-7)
+  [FEATURE] Server: Built the Database Foundation: PostgreSQL, Prisma Schema, and Seed Data